requirements.txt
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
huggingface-hub==0.0.8
joblib==1.0.0
jsonschema==3.2.0
matplotlib==3.3.3
monotonic==1.5
notebook==6.1.6
numpy==1.19.5
oauthlib==3.1.0
packaging==20.8
pandas==1.1.5
pickleshare==0.7.5
prompt-toolkit==3.0.10
ptyprocess==0.7.0
py3nvml==0.2.6
pyarrow==2.0.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
sentencepiece==0.1.91
simplejson==3.17.2
sklearn==0.0
swagger-spec-validator==2.7.3
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.10.3
torch==1.7.1
torchvision==0.8.2
tqdm==4.49.0
transformers==4.6.1
urllib3==1.26.2
xmltodict==0.12.0